Removes field injection and fixes circular dependencies.
